+ [ios] add returnKeyType for textarea and input

+++ ios/sdk/WeexSDK/Sources/Component/WXSliderComponent.m
@@ -607,7 +607,7 @@
         CGFloat width = scrollView.frame.size.width;
         CGFloat XDeviation = scrollView.frame.origin.x - (scrollView.contentOffset.x - width);
         CGFloat offsetXRatio = (XDeviation / width);
-        if (fabsf(offsetXRatio - _lastOffsetXRatio) >= _offsetXAccuracy) {
+        if (fabs(offsetXRatio - _lastOffsetXRatio) >= _offsetXAccuracy) {
             _lastOffsetXRatio = offsetXRatio;
             [self fireEvent:@"scroll" params:@{@"offsetXRatio":[NSNumber numberWithFloat:offsetXRatio]} domChanges:nil];
         }
@@ -662,7 +662,7 @@
     CGFloat width = scrollView.frame.size.width;
     CGFloat XDeviation = scrollView.frame.origin.x - (scrollView.contentOffset.x - width);
     CGFloat offsetXRatio = (XDeviation / width);
-    if (fabsf(offsetXRatio) < 0.5) {
+    if (fabs(offsetXRatio) < 0.5) {
         [self sliderView:self.sliderView scrollViewDidCancelDraging:self.sliderView.scrollView];
     }
 }
